build: add font.js in the repository
The file font.js is generated from font/abc2svg.sfd
by fontforge and base64.
As these programs may not be easily available in some systems,
and as the music font does not evolve frequently, including font.js
in the repository would facilitate the abc2svg generation on
users' machines.